requirements.txt
# Core
pyyaml>=6.0

# LLM providers (install what you need)
openai>=1.30.0
anthropic>=0.25.0

# Utilities
requests>=2.31.0
